ped argument change
ped must now include 4 columns, ID, sire ID, dam ID, and sex. Any additional rows that needed to be added to the pedigree are assumed male. This change is needed to prevent any potential errors when only ID, sire ID, and dam ID are provided and the sex of all animals is assumed male